refactor(habitat): split structural execution lanes
