+++ docs/projects/habitat-harness/deep-refactor/service-shape-backlog.md
@@ -133,3 +133,4 @@ Completed burn-downs:
 - Baseline authority is now its own shared `service/model/baseline` domain with DTO/policy files in named kinds. Check structural reporting and check baseline expansion import baseline explicitly instead of burying baseline authority under `service/model/check`.
 - Source-check scan planning and native source-rule runtime now live under `service/model/source-check`; check no longer re-exports source-check operations, and Nx target inference imports source-check rule module paths from the owning domain.
 - Structural check policy no longer exports source-check helper wrappers; staged source path planning and approved scan-root calculation are owned by `service/model/source-check`.
+- Structural check execution now has named context and source/grit lane policies; the main execution policy orchestrates rule categories instead of also owning provider port shape and source-check execution internals.
